Implemented selection of different tile types base on the type of wall

+++ Procedural Generation Library/Linkage.cpp
@@ -14,7 +14,7 @@ DungeonGenerator* CreateRandomWalkRoom(int width, int height, int startX, int st
 extern "C" __declspec(dllexport)
 int GetSpaceValue(DungeonGenerator* dungeonPtr, int x, int y)
 {
-	return dungeonPtr->GetSpaceValue(x, y);
+	return (int) dungeonPtr->GetSpaceValue(x, y);
 }
 
 extern "C" __declspec(dllexport)
